pyproject.toml
[project]
name = "debt-reversion-ai"
version = "0.1.0"
description = "An AI agent for financial analysis using Dedalus Labs."
requires-python = ">=3.11"
dependencies = [
    "mcp",
    "dedalus-labs",
    "openai==1.100.2",
    "edgartools",
    "yfinance",
    "pandas",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "requests",
    "html2text",
]

[project.scripts]
main = "main:main"

[tool.setuptools]
packages = ["src", "src.servers", "agents"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
